extractor/requirements.txt
typer==0.16.0
pyyaml==6.0.2
toml==0.10.2